+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/algorithm/common/a2c/runner.py
@@ -5,8 +5,8 @@ FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
 """
 
 import numpy as np
-from ch.hslu.wipro.ddpg.new.common.a2c.utils import discount_with_dones
-from ch.hslu.wipro.ddpg.new.common.runners import AbstractEnvRunner
+from ch.hslu.wipro.ddpg.algorithm.common.a2c.utils import discount_with_dones
+from ch.hslu.wipro.ddpg.algorithm.common.runners import AbstractEnvRunner
 
 class Runner(AbstractEnvRunner):
     """
