+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/algorithm/common/models.py
@@ -6,9 +6,9 @@ FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
 
 import numpy as np
 import tensorflow as tf
-from ch.hslu.wipro.ddpg.new.common.a2c import utils
-from ch.hslu.wipro.ddpg.new.common.a2c.utils import conv, fc, conv_to_fc, batch_to_seq, seq_to_batch
-from ch.hslu.wipro.ddpg.new.common.mpi_running_mean_std import RunningMeanStd
+from ch.hslu.wipro.ddpg.algorithm.common.a2c import utils
+from ch.hslu.wipro.ddpg.algorithm.common.a2c.utils import conv, fc, conv_to_fc, batch_to_seq, seq_to_batch
+from ch.hslu.wipro.ddpg.algorithm.common.mpi_running_mean_std import RunningMeanStd
 import tensorflow.contrib.layers as layers
 
 mapping = {}
