+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/algorithm/common/mpi_running_mean_std.py
@@ -9,7 +9,7 @@ try:
 except ImportError:
     MPI = None
 
-import tensorflow as tf, ch.hslu.wipro.ddpg.new.common.tf_util as U, numpy as np
+import tensorflow as tf, ch.hslu.wipro.ddpg.algorithm.common.tf_util as U, numpy as np
 
 class RunningMeanStd(object):
     # https://en.wikipedia.org/wiki/Algorithms_for_calculating_variance#Parallel_algorithm
