+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/algorithm/common/tf_util.py
@@ -292,7 +292,7 @@ def get_placeholder_cached(name):
 # ================================================================
 
 def display_var_info(vars):
-    from ch.hslu.wipro.ddpg.new import logger
+    from ch.hslu.wipro.ddpg.algorithm import logger
     count_params = 0
     for v in vars:
         name = v.name
@@ -318,14 +318,14 @@ def get_available_gpus():
 # ================================================================
 
 def load_state(fname, sess=None):
-    from ch.hslu.wipro.ddpg.new import logger
+    from ch.hslu.wipro.ddpg.algorithm import logger
     logger.warn('load_state method is deprecated, please use load_variables instead')
     sess = sess or get_session()
     saver = tf.train.Saver()
     saver.restore(tf.get_default_session(), fname)
 
 def save_state(fname, sess=None):
-    from ch.hslu.wipro.ddpg.new import logger
+    from ch.hslu.wipro.ddpg.algorithm import logger
     logger.warn('save_state method is deprecated, please use save_variables instead')
     sess = sess or get_session()
     dirname = os.path.dirname(fname)
