+++ Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/algorithm/models.py
@@ -5,7 +5,7 @@ FILE FROM OPEN AI BASELINES https://github.com/openai/baselines
 """
 
 import tensorflow as tf
-from ch.hslu.wipro.ddpg.new.common.models import get_network_builder
+from ch.hslu.wipro.ddpg.algorithm.common.models import get_network_builder
 
 
 class Model(object):
