--- Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/common/__init__.py
@@ -1,5 +1,0 @@
-# flake8: noqa F403
-from ch.hslu.wipro.ddpg.new.common.console_util import *
-from ch.hslu.wipro.ddpg.new.common.dataset import Dataset
-from ch.hslu.wipro.ddpg.new.common.math_util import *
-from ch.hslu.wipro.ddpg.new.common.misc_util import *
