--- Flugsimulator_AI_Landung_Applikation/ch/hslu/wipro/ddpg/new/common/bench/__init__.py
@@ -1,1 +1,0 @@
-from ch.hslu.wipro.ddpg.new.bench.monitor import *
